Add setup instructions to rust game

+++ guessing_game/src/main.rs
@@ -14,7 +14,7 @@ fn main() {
         eprintln!("Strike out! 🏏");
         process::exit(0x0100); // exit with error 256
     }
-    let random_num: u32 = rand::thread_rng().gen_range(1..101);
+    let random_num: u8 = rand::thread_rng().gen_range(1..101);
     println!("Input a number:");
     loop {
         let mut guess = String::new();
@@ -22,7 +22,7 @@ fn main() {
             .read_line(&mut guess)
             .expect("Failed to read line. ");
             println!("You guessed {}", guess);
-            let guess_int: u32 = guess.trim().parse().expect("Error, 🚩 Please input a number! 🚩");
+            let guess_int: u8 = guess.trim().parse().expect("Error, 🚩 Please input a number! 🚩");
             match guess_int.cmp(&random_num) {
                 Ordering::Less => println!("too small 🍓"),
                 Ordering::Greater => println!("too large 🍌"),
